Fix bug where background is black at launch

+++ Source/Utils/Colour.h
@@ -127,9 +127,9 @@ static inline juce::Colour getRainbow12Colour(int value) {
   return juce::Colour::fromFloatRGBA(r, g, b, 1.0f);
 }
 
-static inline juce::ColourGradient getBgGradient(juce::Rectangle<int> boundsRelativeToEditor, float opacity) {
+static inline juce::ColourGradient getBgGradient(juce::Rectangle<int> boundsRelativeToEditor, double opacity) {
   juce::ColourGradient grad = BG_GRADIENT;
-  grad.multiplyOpacity(opacity);
+  grad.multiplyOpacity((float)opacity);
   grad.point1.y = grad.point1.y - static_cast<float>(boundsRelativeToEditor.getY());
   grad.point2.y = grad.point2.y - static_cast<float>(boundsRelativeToEditor.getY());
   return grad;
